Fix a bug where metrics would not auto-update.
Since we changed the metrics to only update if someone is listening, someone
needs to listen or the metrics will not update! However, the charts don't need
to redraw on metric change, since we want the charts to update concurrently on
context change. Now we use a dummy listener to force metrics to update.