Fix race conditions and order dependence.
Metrics are now stored in a single array, rather than creating a temporary copy.
In addition, we make sure that new values are written to the correct place, even
if multiple context change events are received while requests are in-flight.

Metrics now listen to a "beforechange" event which occurs immediately before the
"change" event, allowing the metrics to update their stored values (shifting, if
necessary) before any charts attempt to display them. New values are appended to
the end of the array on a successful "prepare" event, thus avoiding any conflict
or temporary missing data. This also means that the first metric change event
reports times that are consistent with change, rather than with prepare.