Fixed two bugs under FireFox 3.6
I know, I know, FireFox 3.6.  Turns out, some of our developers still use
FF 3.6 to support users that may be on older browsers.  That is, they don't
want to push code that is broken for these users.  This commit should fix
problems with cubeism and FireFox 3.6.  Note: author will need to minify.

First, FireFox 3.6 does not support full ECMA5, which means Object.create()
needed to be stubbed out if it was undefined.

Second, canvas.fillRect() is less forgiving on FF 3.6.  If y1 comes in
undefined, then we should not pass it, or operations based on it, to
fillRect().

This fix tested in latest Safari (MAC), latest FireFox, FireFox 3.6,
and latest Chrome.